Add confimation to Delete

+++ src/components/Delete/Delete.js
@@ -40,18 +40,22 @@ export default function Delete(props){
     )
 }
 
-async function deleteProduct(id){
-    console.log(id)
-    const response = await fetch(`${URL}/products/${id}`, {
-        method: 'DELETE',
-        headers: {
-            'Content-Type': 'application/json'
+async function deleteProduct(id, name){   
+    const confirm = window.confirm(`Are you sure you want to delete ${name}?`);
+
+    if(confirm){
+        const response = await fetch(`${URL}/products/${id}`, {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json'
+            }
+        });
+        console.log(response);
+        if(response.ok){
+            alert("Product Deleted!")
+        } else {
+            alert("Product could not be deleted")
         }
-    });
-    console.log(response);
-    if(response.ok){
-        alert("Product Deleted!")
-    } else {
-        alert("Product could not be deleted")
+        window.location.reload();
     }
 }
